+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3.8'
 services:
   stats-server:
-    build: ewm-stats-service/
+    build: statistics/
     ports:
       - "9090:9090"
     depends_on:
@@ -27,15 +27,15 @@ services:
       - ewm-db
       - stats-server
     environment:
-      MAIN_SPRING_DATASOURCE_URL: jdbc:postgresql://ewm-db:5432/my_main_user
-      MAIN_SPRING_DATASOURCE_USERNAME: my_main_user
-      MAIN_SPRING_DATASOURCE_PASSWORD: my_main_user
-#      STATS_SERVER_URL: http://stats-server:9090
+      MAIN_SPRING_DATASOURCE_URL: jdbc:postgresql://ewm-db:5432/main
+      MAIN_SPRING_DATASOURCE_USERNAME: main
+      MAIN_SPRING_DATASOURCE_PASSWORD: main
+      STATS_SERVER_URL: http://stats-server:9090
 
   ewm-db:
     image: postgres:15.3-alpine3.18
     ports:
       - "6543:5432"
     environment:
-      POSTGRES_USER: my_main_user
-      POSTGRES_PASSWORD: my_main_user
+      POSTGRES_USER: main
+      POSTGRES_PASSWORD: main
