--- ewm-stats-service/stat-client/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-STATS_SERVER_URL: http://stats-server:9090
