+++ pom.xml
@@ -13,7 +13,7 @@
 	<name>Explore With Me</name>
 
 	<modules>
-		<module>ewm-stats-service</module>
+		<module>statistics</module>
 		<module>main-service</module>
 	</modules>
