+++ statistics/pom.xml
@@ -9,7 +9,7 @@
         <version>0.0.1-SNAPSHOT</version>
     </parent>
 
-    <artifactId>ewm-stats-service</artifactId>
+    <artifactId>statistics</artifactId>
     <packaging>pom</packaging>
 
     <modules>
