+++ statistics/stat-dto/src/main/java/ru/practicum/statdto/EndpointHitDto.java
