+++ statistics/stat-service/src/main/java/ru/practicum/statservice/StatServiceApplication.java
