+++ statistics/stat-service/src/main/java/ru/practicum/statservice/controller/StatController.java
