+++ statistics/stat-service/src/main/java/ru/practicum/statservice/exception/ErrorHandler.java
