+++ statistics/stat-service/src/main/java/ru/practicum/statservice/model/EndpointHit.java
