+++ statistics/stat-service/src/main/java/ru/practicum/statservice/model/StatForRequest.java
