+++ statistics/stat-service/src/main/java/ru/practicum/statservice/repository/StatRepository.java
