+++ statistics/stat-service/src/main/java/ru/practicum/statservice/service/StatServiceImpl.java
