+++ statistics/stat-service/src/main/resources/application.properties
