+++ statistics/stat-service/src/main/resources/schema.sql
